Decided on another direction for the game
Instead of recreating the space shooter I decided to make an "Escape the Orbit" game.
Gravity will pull you downwards and you have to try and get high into space.

So I integrated the tilemap from the FlixelEcho example and tried to get collisions to work, but failed. I think it may have to do with adding the body in playstate vs in the class.