Trying ogmo integration
Trying to switch to ogmo for deisgning levels, having some troubles making it work with echo.

Also created a suicider entity that I could use as an enemy (maybe also a follower?), I'm trying to make it get the nearest body to it as a target but I need to understand echo's raycasting first.